Lab4

+++ kernel/printf.c
@@ -137,9 +137,9 @@ void
 backtrace()
 {
   printf("backtrace:\n");
-  uint fp = r_fp();
+  uint64 fp = r_fp();
 
-  uint64 *frame = (uint64 *) fp;
+  uint64 *frame = (uint64*) fp;
   printf("frame[-1]: %p\n", frame[-1]);
   printf("frame[-2]: %p\n", frame[-2]);
   printf("%p\n", fp);
